+++ src/tkc/object_default.c
@@ -150,6 +150,32 @@ static ret_t object_default_get_prop(object_t* obj, const char* name, value_t* v
   return ret;
 }
 
+static bool_t object_default_can_exec(object_t* obj, const char* name, const char* args) {
+  ret_t ret = FALSE;
+  object_default_t* o = OBJECT_DEFAULT(obj);
+  return_value_if_fail(o != NULL, RET_BAD_PARAMS);
+
+  object_t* sub = object_default_get_sub_object(obj, name, &name);
+  if (sub != NULL) {
+    return object_can_exec(sub, name, args);
+  }
+
+  return ret;
+}
+
+static ret_t object_default_exec(object_t* obj, const char* name, const char* args) {
+  ret_t ret = RET_NOT_FOUND;
+  object_default_t* o = OBJECT_DEFAULT(obj);
+  return_value_if_fail(o != NULL, RET_BAD_PARAMS);
+
+  object_t* sub = object_default_get_sub_object(obj, name, &name);
+  if (sub != NULL) {
+    return object_exec(sub, name, args);
+  }
+
+  return ret;
+}
+
 static ret_t object_default_foreach_prop(object_t* obj, tk_visit_t on_prop, void* ctx) {
   ret_t ret = RET_OK;
   object_default_t* o = OBJECT_DEFAULT(obj);
@@ -190,6 +216,8 @@ static const object_vtable_t s_object_default_vtable = {
     .compare = object_default_compare,
     .get_prop = object_default_get_prop,
     .set_prop = object_default_set_prop,
+    .can_exec = object_default_can_exec,
+    .exec = object_default_exec,
     .remove_prop = object_default_remove_prop,
     .foreach_prop = object_default_foreach_prop};